SCI32: fix bug #9752, for QFG4 only
Some SCI32 games have an issue that the saved games thumbnails is useless because it's hidden
by the configuration panel (QFG4) or by the keyboard (Shivers).
Maybe this problems appears on other SCI32 games as well, should be checked.

We avoid that by keeping the thumbnail everytime the menu bar is hidden,
and when time comes to save the game, we use that thumbnail instead of taking
a new one (which would be with the control panel/keyboard).

This commit fixes only QFG4.